extracted <GameBoard/>
